Delete SaveDockerStep
