+++ jib-core/src/test/java/com/google/cloud/tools/jib/builder/steps/SaveAndExtractTarStepTest.java
@@ -17,20 +17,16 @@
 package com.google.cloud.tools.jib.builder.steps;
 
 import com.google.cloud.tools.jib.builder.ProgressEventDispatcher;
-import com.google.cloud.tools.jib.builder.steps.ExtractTarStep.LocalImage;
+import com.google.cloud.tools.jib.builder.steps.SaveAndExtractTarStep.LocalImage;
 import com.google.cloud.tools.jib.cache.Cache;
 import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.event.EventHandlers;
-import com.google.cloud.tools.jib.filesystem.TempDirectoryProvider;
-import com.google.cloud.tools.jib.image.LayerCountMismatchException;
-import com.google.cloud.tools.jib.image.json.BadContainerConfigurationFormatException;
 import com.google.common.io.Resources;
 import com.google.common.util.concurrent.MoreExecutors;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.concurrent.ExecutionException;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
@@ -42,7 +38,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class ExtractTarStepTest {
+public class SaveAndExtractTarStepTest {
 
   @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -70,18 +66,14 @@ public class ExtractTarStepTest {
   }
 
   @Test
-  public void testCall_validDocker()
-      throws URISyntaxException, LayerCountMismatchException,
-          BadContainerConfigurationFormatException, IOException, ExecutionException,
-          InterruptedException {
+  public void testCall_validDocker() throws Exception {
     Path dockerBuild = getResource("core/extraction/docker-save.tar");
     LocalImage result =
-        new ExtractTarStep(
+        SaveAndExtractTarStep.tarImageStep(
                 MoreExecutors.newDirectExecutorService(),
                 buildConfiguration,
-                dockerBuild,
                 progressEventDispatcherFactory,
-                new TempDirectoryProvider())
+                dockerBuild)
             .call();
 
     Mockito.verify(progressEventDispatcher, Mockito.times(2)).newChildProducer();
@@ -102,18 +94,14 @@ public class ExtractTarStepTest {
   }
 
   @Test
-  public void testCall_validTar()
-      throws URISyntaxException, LayerCountMismatchException,
-          BadContainerConfigurationFormatException, IOException, ExecutionException,
-          InterruptedException {
+  public void testCall_validTar() throws Exception {
     Path tarBuild = getResource("core/extraction/jib-image.tar");
     LocalImage result =
-        new ExtractTarStep(
+        SaveAndExtractTarStep.tarImageStep(
                 MoreExecutors.newDirectExecutorService(),
                 buildConfiguration,
-                tarBuild,
                 progressEventDispatcherFactory,
-                new TempDirectoryProvider())
+                tarBuild)
             .call();
 
     Mockito.verify(progressEventDispatcher, Mockito.times(2)).newChildProducer();
@@ -135,7 +123,9 @@ public class ExtractTarStepTest {
 
   @Test
   public void testIsGzipped() throws URISyntaxException, IOException {
-    Assert.assertTrue(ExtractTarStep.isGzipped(getResource("core/extraction/compressed.tar.gz")));
-    Assert.assertFalse(ExtractTarStep.isGzipped(getResource("core/extraction/not-compressed.tar")));
+    Assert.assertTrue(
+        SaveAndExtractTarStep.isGzipped(getResource("core/extraction/compressed.tar.gz")));
+    Assert.assertFalse(
+        SaveAndExtractTarStep.isGzipped(getResource("core/extraction/not-compressed.tar")));
   }
 }
