+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/JibBuildRunner.java
@@ -272,9 +272,12 @@ public class JibBuildRunner {
       throw new BuildStepsExecutionException(message, ex);
 
     } catch (ExecutionException ex) {
-      String message = Verify.verifyNotNull(ex.getCause().getMessage()); // keep null-away happy
-      throw new BuildStepsExecutionException(message, ex.getCause());
-
+      if (ex.getCause().getMessage() != null) {
+        String message = Verify.verifyNotNull(ex.getCause().getMessage()); // keep null-away happy
+        throw new BuildStepsExecutionException(message, ex.getCause());
+      } else {
+        throw new BuildStepsExecutionException(Verify.verifyNotNull(ex.getMessage()), ex);
+      }
     } catch (InterruptedException ex) {
       throw new BuildStepsExecutionException(helpfulSuggestions.none(), ex);
     }
